# apps/server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Triathlon API",
    "version": "0.2.0",
    "description": "One public HTTP interface for all Triathlon clients. Error responses share the Problem Details envelope with a stable domain code."
  },
  "components": {
    "securitySchemes": {
      "bearerKey": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/projects": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "timezone": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "name",
                    "timezone",
                    "createdBy",
                    "ownerUserId",
                    "createdAt",
                    "deletedAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "timezone": {
                      "type": "string"
                    },
                    "createdBy": {
                      "type": "string"
                    },
                    "ownerUserId": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "deletedAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "project",
                      "role"
                    ],
                    "properties": {
                      "project": {
                        "type": "object",
                        "required": [
                          "id",
                          "name",
                          "timezone",
                          "createdBy",
                          "ownerUserId",
                          "createdAt",
                          "deletedAt"
                        ],
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "timezone": {
                            "type": "string"
                          },
                          "createdBy": {
                            "type": "string"
                          },
                          "ownerUserId": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "deletedAt": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          }
                        }
                      },
                      "role": {
                        "anyOf": [
                          {
                            "type": "string",
                            "enum": [
                              "owner"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "member"
                            ]
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "projectId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "project",
                    "role",
                    "columns",
                    "members"
                  ],
                  "properties": {
                    "project": {
                      "type": "object",
                      "required": [
                        "id",
                        "name",
                        "timezone",
                        "createdBy",
                        "ownerUserId",
                        "createdAt",
                        "deletedAt"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "timezone": {
                          "type": "string"
                        },
                        "createdBy": {
                          "type": "string"
                        },
                        "ownerUserId": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "deletedAt": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      }
                    },
                    "role": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "owner"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "member"
                          ]
                        }
                      ]
                    },
                    "columns": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "projectId",
                          "name",
                          "category",
                          "position",
                          "createdAt"
                        ],
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "projectId": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "category": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "not_started"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "started"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "done"
                                ]
                              }
                            ]
                          },
                          "position": {
                            "type": "integer"
                          },
                          "createdAt": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "members": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "userId",
                          "joinedAt",
                          "role"
                        ],
                        "properties": {
                          "userId": {
                            "type": "string"
                          },
                          "joinedAt": {
                            "type": "string"
                          },
                          "role": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "owner"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "member"
                                ]
                              }
                            ]
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "timezone": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "projectId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "name",
                    "timezone",
                    "createdBy",
                    "ownerUserId",
                    "createdAt",
                    "deletedAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "timezone": {
                      "type": "string"
                    },
                    "createdBy": {
                      "type": "string"
                    },
                    "ownerUserId": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "deletedAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "projectId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "name",
                    "timezone",
                    "createdBy",
                    "ownerUserId",
                    "createdAt",
                    "deletedAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "timezone": {
                      "type": "string"
                    },
                    "createdBy": {
                      "type": "string"
                    },
                    "ownerUserId": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "deletedAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/transfer": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "toUserId"
                ],
                "properties": {
                  "toUserId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "projectId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "name",
                    "timezone",
                    "createdBy",
                    "ownerUserId",
                    "createdAt",
                    "deletedAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "timezone": {
                      "type": "string"
                    },
                    "createdBy": {
                      "type": "string"
                    },
                    "ownerUserId": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "deletedAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/recover-ownership": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "toUserId"
                ],
                "properties": {
                  "toUserId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "projectId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "name",
                    "timezone",
                    "createdBy",
                    "ownerUserId",
                    "createdAt",
                    "deletedAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "timezone": {
                      "type": "string"
                    },
                    "createdBy": {
                      "type": "string"
                    },
                    "ownerUserId": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "deletedAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{projectId}/purge": {
      "delete": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "projectId",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://127.0.0.1:1"
    }
  ]
}